pcd: move the proof corruption accessors out of proof/mod.rs
The mutable component accessors the corruption vocabulary needs were a
111-line `#[cfg(feature = "unstable-fuzzing")]` block sitting in the middle
of the production proof module. Feature-gated is not the same as isolated:
anyone reading `proof/mod.rs` still had to page past a fuzzing-only API.

They move to `proof::fuzz`, a child module compiled only under the feature.
Being a child of `proof` is what keeps this possible without loosening
anything — the `Cached` bridge polynomials and commitment caches stay
private to `proof`, and a descendant module can still reach them. What is
left behind is the `mod` declaration.
